Refactored file locations
