+++ lib/pages/catalog/product/home/product.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
-import 'package:exdock_backend_client/pages/catalog/products/product_data.dart';
-import 'package:exdock_backend_client/pages/catalog/products/product_home_synchronous.dart';
+import 'package:exdock_backend_client/pages/catalog/product/home/product_data.dart';
+import 'package:exdock_backend_client/pages/catalog/product/home/product_home_synchronous.dart';
 import 'package:flutter/material.dart';
 
 class Product extends StatelessWidget {
