+++ lib/pages/catalog/product/home/product_home_synchronous.dart
@@ -1,7 +1,7 @@
-import 'package:exdock_backend_client/pages/catalog/products/product_data.dart';
-import 'package:exdock_backend_client/pages/catalog/products/widgets/middle_bar/product_middle_bar.dart';
-import 'package:exdock_backend_client/pages/catalog/products/widgets/product_list/product_list.dart';
-import 'package:exdock_backend_client/pages/catalog/products/widgets/top_bar/product_top_bar.dart';
+import 'package:exdock_backend_client/pages/catalog/product/home/product_data.dart';
+import 'package:exdock_backend_client/pages/catalog/product/home/widgets/middle_bar/product_middle_bar.dart';
+import 'package:exdock_backend_client/pages/catalog/product/home/widgets/product_list/product_list.dart';
+import 'package:exdock_backend_client/pages/catalog/product/home/widgets/top_bar/product_top_bar.dart';
 import 'package:flutter/material.dart';
 
 class ProductHomeSynchronous extends StatefulWidget {
