+++ lib/pages/catalog/product/home/widgets/middle_bar/product_view_size.dart
