+++ lib/pages/catalog/product/home/widgets/middle_bar/remove_filter_button.dart
