+++ lib/pages/catalog/product/home/widgets/product_list/product_info.dart
