+++ lib/router/routes/catalog.dart
@@ -1,5 +1,5 @@
 import 'package:exdock_backend_client/pages/catalog/category/category.dart';
-import 'package:exdock_backend_client/pages/catalog/products/product.dart';
+import 'package:exdock_backend_client/pages/catalog/product/home/product.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
@@ -12,7 +12,7 @@ List<GoRoute> getCatalogRoutes() {
       builder: (context, state) => const Placeholder(),
     ),
     GoRoute(
-      path: '/catalog/products',
+      path: '/catalog/product',
       builder: (context, state) => Product(),
     ),
     GoRoute(
@@ -20,10 +20,11 @@ List<GoRoute> getCatalogRoutes() {
       builder: (context, state) => Category(),
     ),
     GoRoute(
-        path: '/catalog/product/:id',
-        builder: (context, state) {
-          final id = state.pathParameters['id'];
-          return ProductInfo();
-        }),
+      path: '/catalog/product/:id',
+      builder: (context, state) {
+        final id = state.pathParameters['id'];
+        return ProductInfo();
+      },
+    ),
   ];
 }
